Fixing Accessibility Bug in WPF Sample(#5403)
Description
ExpenseItIntro-> the "view expense report" title and the background color contrast is failing.
ExpenseItIntro->Expense report for-Color contrast is failing for table and overall background.
ExpenseItIntro->Expense report for-Color contrast for letters in the table and the blue background is failing.
DataBindingDemo->List of products-color contrast for the given highlighted colors is not compliant to the existing ratio.